requirements.txt
Flask
flask-cors
opencv-python-headless
numpy
Pillow
tensorflow==2.12.0
keras
gTTS
gunicorn
